dom/indexedDB/test/test_remove_objectStore.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<html>
<head>
  <title>Indexed Database Property Test</title>

  <script type="text/javascript" src="/MochiKit/packed.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>

  <script type="text/javascript">
    function run_test()
    {
      const nsIIDBObjectStore = Components.interfaces.nsIIDBObjectStore;
      const name = window.location.pathname;
      const description = "My Test Database";
      const objectStoreName = "Objects";

      function errorHandler(event) {
        ok(false, "indexedDB error (" + event.error.code + "): " +
           event.error.message);
        SimpleTest.finish();
      }

      var db = indexedDB.open(name, description);

      var count = db.objectStores.length;
      is(count, 0, "Bad objectStores list");

      var request = db.createObjectStore(objectStoreName, "foo");
      request.onerror = errorHandler;
      request.onsuccess = function(event) {
        is(db.objectStores.length, 1, "Bad objectStores list");
        request = db.openObjectStore(objectStoreName,
                                     nsIIDBObjectStore.READ_ONLY);
        request.onerror = errorHandler;
        request.onsuccess = function(event) {
          var objectStore = event.result;
          is(objectStore.name, objectStoreName, "Bad name");
          is(objectStore.mode, nsIIDBObjectStore.READ_ONLY, "Bad mode");
          is(objectStore.keyPath, "foo", "Bad keyPath");
          if(objectStore.indexNames.length, 0, "Bad indexNames");
          is(db.objectStores.length, 1, "Bad objectStores list");
          request = db.removeObjectStore(objectStoreName);
          request.onerror = errorHandler;
          request.onsuccess = function(event) {
            is(event.result, undefined, "Bad result");
            is(db.objectStores.length, 0, "Bad objectStores list");
            SimpleTest.finish();
          };
        };
      };
    }

    SimpleTest.waitForExplicitFinish();
  </script>

</head>

<body onload="run_test();">
</body>
</html>
